pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "subtitle-tool"
version = "0.2.0"
description = "Desktop tool for Japanese transcription, English subtitle generation, and context-guided subtitle rewrites."
readme = "README.md"
requires-python = ">=3.11,<3.12"
license = { text = "MIT" }
dependencies = [
  "accelerate>=1.0",
  "psutil>=5.9",
  "PyYAML>=6.0",
  "requests>=2.32",
  "safetensors>=0.4",
  "tomli-w>=1.2",
  "transformers>=4.46",
]

[project.optional-dependencies]
dev = [
  "Pillow>=11.0",
  "pytest>=8.3",
]
gpu = [
  "torch>=2.6",
]
build = [
  "pyinstaller>=6.16",
]

[project.scripts]
subtitle-tool = "local_subtitle_stack.launcher:main"
subtitle-stack = "local_subtitle_stack.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
